fix: User.householdSize 타입을 SMALLINT에 맞춰 Short로 수정[#23]
household_size 컬럼이 SMALLINT인데 엔티티는 Integer로 매핑되어 있어
Hibernate 스키마 검증에 실패했다. ChecklistItem.orderNo와 같은
원인이라 함께 수정한다.

+++ src/main/java/com/sayyoung/seed/domain/user/entity/User.java
@@ -54,7 +54,7 @@ public class User {
     private Boolean basicRecipient;
 
     @Column(name = "household_size")
-    private Integer householdSize;
+    private Short householdSize;
 
     @Column(name = "budget")
     private Long budget;
@@ -90,7 +90,7 @@ public class User {
             String regionCode,
             Long income,
             Boolean basicRecipient,
-            Integer householdSize
+            Short householdSize
     ) {
         this.protectionEndDate = protectionEndDate;
         this.youthSupport = youthSupport;
